requirements.txt
Flask
cryptography
scikit-fuzzy
python-aiml
http://127.0.0.1:5000/